requirements.txt
python=3.11.9
imageio=2.33.1
matplotlib=3.9.2
matplotlib-base=3.9.2
numpy=1.26.4
numpy-base=1.26.4
pandas=2.2.2
pytorch=2.3.0
scikit-learn=1.5.1
scipy=1.13.1
torchvision=0.18.1
